Homework extended done

+++ src/main/java/com/eshop/controller/ProductController.java
@@ -2,7 +2,12 @@ package com.eshop.controller;
 
 
 import com.eshop.model.Product;
+import com.eshop.model.Role;
+import com.eshop.model.User;
+import com.eshop.model.exceptions.NotYourProductException;
 import com.eshop.model.web.ProductRequest;
+import com.eshop.model.web.ProductUpdateRequest;
+import com.eshop.repository.UserSessionRepository;
 import com.eshop.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -16,9 +21,20 @@ public class ProductController {
     @Autowired
     private ProductService productService;
 
+    @Autowired
+    private UserSessionRepository userSessionRepository;
+
     @PostMapping("/create")
-    public Product create(@RequestBody ProductRequest productRequest) {
-        return productService.create(productRequest);
+    public Product create(@RequestBody ProductRequest productRequest,
+                          @RequestHeader("Authorization") String sessionId) {
+        return productService.create(productRequest, sessionId);
+    }
+
+    @PutMapping("/update/{productId}")
+    public Product update(@RequestBody ProductUpdateRequest productUpdateRequest,
+                          @PathVariable("productId") Integer productId,
+                          @RequestHeader("Authorization") String sessionId){
+        return productService.update(productUpdateRequest, productId, sessionId);
     }
 
     @GetMapping("/all")
@@ -37,8 +53,8 @@ public class ProductController {
     }
 
     @GetMapping("/search/seller")
-    public List<Product> searchBySellerId(@RequestParam("sellerid") Integer sellerId) {
-        return productService.findBySellerId(sellerId);
+    public List<Product> searchBySellerId(@RequestParam("username") String username) {
+        return productService.findBySeller(username);
     }
 
     @GetMapping("/search/price")
@@ -52,9 +68,21 @@ public class ProductController {
         return productService.findByCategory(categoryId);
     }
 
+    @GetMapping("/searchParams")
+    public List<Product> search(@RequestParam(value = "name", required = false) String productName,
+                                @RequestParam(value = "catid", required = false) Integer categoryId,
+                                @RequestParam(value = "isDescending", required = false, defaultValue = "true") Boolean isDesc) {
+        return productService.findByParamsy(productName, categoryId,isDesc);
+    }
+
     @DeleteMapping("/delete/{productId}")
-    public void deleteByProductId(@PathVariable("productId") Integer productId){
-        productService.delete(productId);
+    public void deleteByProductId(@PathVariable("productId") Integer productId, @RequestHeader("Authorization") String sessionId) throws NotYourProductException {
+        User user = userSessionRepository.findBySessionId(sessionId).getUser();
+        if (productService.findById(productId).getSeller().equals(user) || user.getRoles().contains(Role.ADMIN)) {
+            productService.delete(productId);
+        } else {
+            throw new NotYourProductException("Only seller can delete his product");
+        }
     }
 
 }
